backend/poll_iot_messages.py
#!/usr/bin/env python3
"""
Poll AWS IoT Core for recent messages and update backend data store
This is a workaround until we set up IoT Shadow or IoT Rules
"""
import boto3
import json
import os
from datetime import datetime, timedelta
from dotenv import load_dotenv

load_dotenv()

# This simulates getting data from IoT
# In production, use IoT Shadow or IoT Rules → Lambda

def get_latest_from_mqtt_test():
    """
    Note: AWS IoT Core doesn't have a direct API to query recent MQTT messages.
    Options:
    1. Use IoT Shadow (ESP32 updates shadow)
    2. Use IoT Rules → DynamoDB → Query DynamoDB
    3. Use IoT Rules → Lambda → Backend API
    4. Use MQTT subscription (iot_subscriber.py)
    
    For now, this is a placeholder that shows what needs to be done.
    """
    print("=" * 60)
    print("IoT Message Polling")
    print("=" * 60)
    print()
    print("⚠️  AWS IoT Core doesn't provide an API to query recent MQTT messages.")
    print()
    print("To get real-time ESP32 data, use one of these approaches:")
    print()
    print("1. IoT Shadow (Recommended):")
    print("   - ESP32 updates shadow with sensor data")
    print("   - Backend queries shadow (already implemented)")
    print()
    print("2. IoT Rules → DynamoDB:")
    print("   - Create IoT Rule that forwards messages to DynamoDB")
    print("   - Backend queries DynamoDB for latest data")
    print()
    print("3. IoT Rules → Lambda → Backend:")
    print("   - IoT Rule triggers Lambda on new messages")
    print("   - Lambda calls backend API to update data")
    print()
    print("4. MQTT Subscription:")
    print("   - Run iot_subscriber.py in background")
    print("   - Subscribes to topics and updates data store")
    print()
    print("Current status: Backend is querying shadow, but ESP32")
    print("needs to update shadow for data to appear.")

if __name__ == '__main__':
    get_latest_from_mqtt_test()




